initial 1.24 k8s support plus go1.18
