+++ pkg/prober/prober.go
@@ -3,6 +3,7 @@ package prober
 import (
 	"crypto/tls"
 	"crypto/x509"
+	"fmt"
 	"io/ioutil"
 	"net/http"
 	"net/url"
@@ -136,4 +137,18 @@ func DoProbe(probe Probe, probeStatus *ProbeStatus, initial bool) error {
 	}
 
 	return nil
+}
+
+// GetSystemCertPool returns a x509.CertPool that contains the
+// root CA certificates if they are present at runtime
+func GetSystemCertPool(probeName string) (*x509.CertPool, error) {
+	caCertPool, err := x509.SystemCertPool()
+	if err != nil {
+		caCertPool = x509.NewCertPool()
+		logrus.Errorf("[GetSystemCertPoolUnix] error loading system cert pool for probe (%s): %v", probeName, err)
+	}
+	if caCertPool == nil {
+		return nil, fmt.Errorf("[GetSystemCertPoolWindows] x509 returned a nil certpool for probe (%s)", probeName)
+	}
+	return caCertPool, nil
 }
